Broke out Search combobox to individual fubnctional component.
