+++ client/components/home-map.jsx
@@ -40,6 +40,7 @@
 // }
 
 import React, { useContext } from 'react';
+import Search from './search';
 import HomeContext from '../lib/home-context';
 
 import {
@@ -48,20 +49,6 @@ import {
   Marker,
   InfoWindow
 } from '@react-google-maps/api';
-
-import usePlacesAutocomplete, {
-  getGeocode,
-  getLatLng
-} from 'use-places-autocomplete';
-
-import {
-  Combobox,
-  ComboboxInput,
-  ComboboxPopover,
-  ComboboxList,
-  ComboboxOption
-} from '@reach/combobox';
-import '@reach/combobox/styles.css';
 
 import mapStyles from '../lib/map-styles';
 
@@ -154,58 +141,5 @@ export default function HomeMap() {
       </GoogleMap>
     </div>
 
-  );
-}
-
-function Search({ panTo }) {
-  const {
-    ready,
-    value,
-    suggestions: { status, data },
-    setValue,
-    clearSuggestions
-  } = usePlacesAutocomplete({
-    requestOptions: {
-      location: { lat: () => 43.6532, lng: () => -79.3832 },
-      radius: 100 * 1000
-    }
-  });
-
-  const handleInput = e => {
-    setValue(e.target.value);
-  };
-
-  const handleSelect = async address => {
-    setValue(address, false);
-    clearSuggestions();
-
-    try {
-      const results = await getGeocode({ address });
-      const { lat, lng } = await getLatLng(results[0]);
-      panTo({ lat, lng });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  return (
-    <div className="search">
-      <Combobox onSelect={handleSelect}>
-        <ComboboxInput
-          value={value}
-          onChange={handleInput}
-          disabled={!ready}
-          placeholder="Search your location"
-        />
-        <ComboboxPopover>
-          <ComboboxList>
-            {status === 'OK' &&
-              data.map(({ id, description }) => (
-                <ComboboxOption key={id} value={description} />
-              ))}
-          </ComboboxList>
-        </ComboboxPopover>
-      </Combobox>
-    </div>
   );
 }
